requirements.txt
asgiref==3.8.1
astroid==3.2.2
attrs==24.2.0
beautifulsoup4==4.12.3
black==24.4.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dill==0.3.8
distlib==0.3.8
filelock==3.15.4
fonttools==4.53.1
h11==0.14.0
idna==3.10
isort==5.13.2
kiwisolver==1.4.5
matplotlib==3.9.1
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==2.0.0
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
pycparser==2.22
pylint==3.2.3
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==308
requests==2.32.3
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.0
tomlkit==0.12.5
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
virtualenv==20.26.3
virtualenvwrapper-win==1.2.7
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
